+++ src/main/java/lotto/domain/LottoSettings.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public enum LottoSettings {
     MIN(1),
