+++ src/main/java/lotto/domain/Money.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Objects;
