+++ src/main/java/lotto/domain/UniqueRandomNumberProvider.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Collections;
 import java.util.List;
